pyproject.toml
[tool.poetry]
name = "autopack-demo"
version = "0.1.0"
description = ""
authors = ["Erik Peterson <e@eriklp.com>"]
packages = [{ include = "demo" }]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.2"
langchain = "^0.0.215"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
psutil = "^5.9.5"
autopack-tools = "^0.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
